Forms: move field-icons.tsx to response-fields directory
